Revert "Changed the icons and also included the background"
