+++ package.json
@@ -4,16 +4,15 @@
   "private": true,
   "dependencies": {
     "@paystack/inline-js": "^2.12.6",
-    "@react-md/icon": "^5.1.3",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-icons": "^4.8.0",
+    "react-icons": "^4.7.1",
     "react-lazy-load-image-component": "^1.5.6",
-    "react-router-dom": "^6.10.0",
+    "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
     "react-select": "^5.7.0",
     "web-vitals": "^2.1.4"
