+++ src/components/navbar/navbar-component.css
@@ -61,7 +61,7 @@
   }
   .each_link{
     margin: 3rem 2.5rem;
-    color: white;border:2px solid red
+    color: white;
   }
 
   .nav_toggle {
