+++ src/components/testimony/testimony.css
@@ -37,7 +37,6 @@
 .testimony aside aside div span p {
   font-size: 12px;
   line-height: 18px;
-  margin:0;
 }
 .testimony aside aside > p {
   font-size: 13px;
@@ -73,7 +72,7 @@
     line-height: 30px;
   }
   .testimony h3 {
-    font-size: 40px;
+    font-size: 35px;
     line-height: 60px;
     margin: 0;
   }
