+++ src/components/verified-orphanages/verified-orphanages.css
@@ -49,7 +49,6 @@
   .verified-orphanages h3 {
     font-size: 40px;
     line-height: 60px;
-    font-weight:700;
   }
   .verified-orphanages .mobileOrphanage {
     display: none;
@@ -69,11 +68,9 @@
   .verified-orphanages aside div h4 {
     font-size: 24px;
     line-height: 36px;
-    font-weight:700;
   }
   .verified-orphanages aside div p {
     font-size: 16px;
     line-height: 24px;
-    font-weight:500;
   }
 }
