+++ src/components/why-us-component/why-us-component.css
@@ -19,8 +19,8 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 100px;
-  height: 100px;
+  width: 5rem;
+  height: 100%;
   border-radius: 50%;
   margin-bottom: 20px;
   margin-top: 50px;
@@ -31,7 +31,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 48px;
+  width: 100%;
   height: 100%;
 }
 .why-us div .image-wrapper span .icon {
