+++ src/pages/each-orphanages-page/EachOrphanages.js
@@ -1,12 +1,9 @@
 import Campaign from "../../components/campaigns/campaigns";
 import Each_home_page from "../../components/each-home-page/eachHomePage";
-import OurMissionComponent from "../../components/our-mission-component/our-mission-component";
 import Media from "../../components/media/Media";
 import Navbar from "../../components/navbar/navbar-component";
 import PersuasionPage from "../../components/persuasion-page/persuasion-page";
-import img from '../../assets/orphanages-photos/img-7.png';
 import { Helmet } from "react-helmet";
-
 
 const EachOrphanages = () => {
   const message =
@@ -27,7 +24,6 @@ const EachOrphanages = () => {
       />
       <div>
         <Each_home_page />
-        <OurMissionComponent img={img} />
         <Campaign />
         <Media />
         <PersuasionPage message={message} />
